Bank-Simulator/Bank_Simulator/requirements.txt
cffi==1.17.1
cryptography==44.0.2
mysql-connector==2.2.9
pycparser==2.22
tabulate==0.9.0
